Build a form using react-hook-form
Able to take user input for game configuration and package it into one nice JS object to be sent to server. A lot of form validation comes built-in by HTML and enforced by browser. Though now will be tricky to figure out how to navigate on button click with React router